requirements.txt
discord.py
python-dotenv
datetime
openai
ollama
langdetect